requirements/base.txt
fastapi==0.86.0
psycopg2==2.9.5
uvicorn==0.19.0
yoyo-migrations==8.1.0
